Profile data endpoints

+++ src/TTRPGBE-1/src/main/java/com/ttrpg/repository/KorisnikRepository.java
@@ -1,7 +1,6 @@
 package com.ttrpg.repository;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -10,7 +9,7 @@ import com.ttrpg.model.Korisnik;
 
 @Repository
 
-public interface KorisnikRepository extends JpaRepository <Korisnik,Long>{
+public interface KorisnikRepository extends JpaRepository <Korisnik,Integer>{
 	
 	
      List<Korisnik> findByUsername(String username);
